Workflows for basic pathogen repo CI
The goal is to establish a common setup for CI for all or at least most
of our pathogen build repos.

Includes a reusable workflow¹ which can be called/invoked by our
workflows in other repos and a "starter" workflow template² that
performs such a call, for copying into the pathogen repo (either
manually or via the web UI which supports these starter workflows).

These are both basic for now but appropriate for our "simple" pathogen
builds like zika and measles.  We can add workflow parameters and
conditionals as necessary in the future to support CI for more complex
builds, like ncov.  (It won't take much!)

Related to <https://github.com/nextstrain/zika/pull/16>.

¹ https://docs.github.com/en/actions/using-workflows/reusing-workflows
² https://docs.github.com/en/actions/using-workflows/creating-starter-workflows-for-your-organization