pathogen-repo-ci: Allow caller to pass env vars
This can be useful to set variables respected by Augur, Nextstrain CLI,
etc.  Implemented ourselves as a workflow input since the usual job.env
block is not (yet?) supported by GitHub Actions for jobs which call a
reusable workflow.

The input value is a YAML-formatted string, which you can easily produce
by pretending you're writing normal nested YAML within a literal
multi-line block scalar (introduced by "|"):

   with:
     env: |
       FOO: bar
       I_CANT_BELIEVE: "it's not YAML"

When (if?) the usual job.env block is supported for reusable workflows,
it will be a small transformation for each caller to use it.