sync-rtd-redirects: Skip when RTD_TOKEN is empty
RTD redirects can only be synced with a valid RTD_TOKEN. If it is an
empty string, it must be invalid. Ideally, making the secret required
would catch all invalid token errors, however an empty string still
passes the requirement. This job-level condition is the next best thing.

When called with an empty token, the workflow will still be triggered as
configured but the job will be skipped instead of failing due to an
invalid RTD token.