Workflow template to launch a runner for ad-hoc interactive debugging…
…using our setup-ssh action in the nextstrain/.github repo.

Useful to test changes to setup-ssh, but also useful to have available
for debugging runners outside of the context of any specific workflow
(i.e. in which you could use setup-ssh directly).

Provided as a workflow template, not an actual runnable workflow in this
repository, because this workflow should only be used in private
repositories (which this repo is not).  Instead, see our
nextstrain/private repo for a runnable copy of this workflow.¹

¹ https://github.com/nextstrain/private/actions/workflows/debugging-runner.yaml