actions/setup-debugger: Interactively inspect and manipulate steps of a workflow job
Interception is still slightly flaky at times for as-of-yet unclear
reasons, but overall it works!  This is a quick initial implementation
to see if such a debugger is actually useful in practice.  There are
alternative ways to accomplish the ~same functionality, which I'll
describe elsewhere later (and one of which I've prototyped).  One of
those may ultimately be a better choice if this debugger does prove its
worth.