actions/shellcheck: Run ShellCheck on all shell programs in a repo
Automatically finds all shell programs that can be checked, including
snippets in GitHub Actions workflows.  We can use this action to
standardize on the use of ShellCheck as a linter for our repos.

The shellcheck wrapper program and extract-shell-from-gh-workflow
program are based on the programs of the same name that I wrote for
ncov-ingest.¹  I've updated them here with additional functionality for
more generalized use.

¹ <https://github.com/nextstrain/ncov-ingest/blob/3b711bf4/bin/shellcheck>
  <https://github.com/nextstrain/ncov-ingest/blob/3b711bf4/bin/extract-shell-from-gh-workflow>