Add actions/run-nextstrain-build-on-aws-batch
Creates a custom action to run Nextstrain builds on AWS Batch so
that we can centralize and standardize how our pathogen repos set up
these builds. Motivated by the latest hiccup with the Snakemake upgrade
that required fixes across multiple pathogen repos.¹

I decided to make it a composite action instead of a reusable workflow
to allow for per-repo custom steps such as setting up trial configs
for trial builds² and to be able to use the `GITHUB_ACTION_PATH` to
access the shared write-envdir script.

¹ https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1681328249848399
² https://github.com/nextstrain/ncov/blob/4c5fc8f484cb7310b1039f5f8e847dad34fe4774/.github/workflows/rebuild-gisaid.yml#L38-L48