pathogen-repo-ci: Skip logging into docker.io when the org-level secret isn't available
Even though a login that fails due to the missing secret doesn't affect
the workflow's execution status, it does clutter the workflow's summary
page with a bunch of noisy warnings.

Suggestion to do this pre-check from @joverlee521 in review.  If only we
could check for secrets directly in a step condition ("if: …").