Add bin/interpolate-text-template
Modifies the "Generate summary" step from the conda-base repo CI
workflow¹ into a separate script that can be used to interpolate any
text template with environment variables.

I will be using this in the reusable run-build workflow to create the
GitHub Action summary for AWS Batch builds. I also see this being useful
for creating Slack messages for automated builds such as the onstart
and onerror notifications that we use in ncov² if we decide to move the
notifications out of the pathogen workflows and into the GitHub Action
workflows.

¹ https://github.com/nextstrain/conda-base/blob/5655133be6864ede0acbf40b6b06222f9f2dc031/.github/workflows/ci.yaml#L107-L112
² https://github.com/nextstrain/ncov/blob/fcac1d16410b1b119a9af70afe7b9325a6e25083/workflow/snakemake_rules/export_for_nextstrain.smk#L499-L519