Add bin/json-to-envvars and bin/yaml-to-envvars
Modifies the "Set environment variables" step from the pathogen-repo-ci
workflow¹ into two separate scripts since we will use the same method to
set environment variables to other reusable workflows.

I decided to create `json-to-envvars` as a separate script that
`yaml-to-envvars` calls because I will be using it to set secrets as
environment variables as well. The `--varnames` option will allow us to
be explicit about which secrets we want to set as environment variables.

¹ https://github.com/nextstrain/.github/blob/cc6f4385a45bd6ed114ab4840416fd90cc46cd1b/.github/workflows/pathogen-repo-ci.yaml#L183-L198