bin/json-to-envvars: convert `null` to empty strings
Convert the null values to empty strings so that we do not set the
value "null" as any environment variable values.

I had considered just skipping key/value pairs where the value is null,
but I think I'd want to see that the environment variable is empty in
the GitHub Action logs.

This change was motivated by a failed test seasonal-flu build on AWS
Batch. The workflow passed `NEXTSTRAIN_DOCKER_IMAGE: null` to the
Nextstrain CLI and the AWS Batch job tried to pull a `null` image.
With this change, the Nextstrain CLI is able to see that the environment
variable as a falsy value and the AWS Batch job pulled the default
latest image.