attach-aws-batch: Add AWS Batch console URL
Include a direct URL to the AWS Batch console in our AWS Batch summary
as requested by @jameshadfield¹

Note: in order for the URL in the summary to _not_ be masked, the repo
cannot have the `us-east-1` value as a secret. This means we would need
to audit repos of the `AWS_DEFAULT_REGION` secret and remove them where
possible.

¹ https://github.com/nextstrain/.github/pull/39#discussion_r1208660848