run-build: Use `workflow-context`
Replace the `ref` input with our new `workflow-context` composite action
to get the repository and sha of the called reusable workflow of
the current run.

This removes the burden of having to provide the ref as an input by
the caller workflow. However, the caller workflow does need to provide
the permissions `id-token: write` in order to use this workflow. This
has been reflected in the workflow-template.