pathogen-repo-build: Prefix artifact-paths with `BUILD_DIR`
The user will be defining artifact paths based on their pathogen repo
paths, but we checkout the pathogen repo under the `BUILD_DIR`. In order
for the artifact-paths to match, we must prefix the provided path with
the `BUILD_DIR`.

Also adds the `BUILD_DIR` prefix to the excluded `ENV_DIR` path since
we create it within the build directory.