pathogen-repo-build: Explicitly define accepted secrets
Be explicit about which secrets are being passed to the workflow, which
will open the way for us to just pass all environment variables to
the build runtime without the `runtime-envvar-names` input.

The main tradeoffs of this approach as opposed to using `secrets:inherit`:

1. The calling workflows must now list out all of the secrets to pass
to the reusable workflow.

2. If any pathogen builds require additional secrets in the future, we
will have to update this reusable workflow to accept the new secrets.