pathogen-repo-build: remove `runtime-envvar-names`
Instead of asking calling workflows to provide the
`runtime-envvar-names` input, just pass all explicit envvars and secrets
to the build runtime.

Added bogus `env` and `secrets` to the CI to ensure that we are testing
the addition of build runtime environment variables.

This change is motivated by @jameshadfield's review comment
https://github.com/nextstrain/.github/pull/39#discussion_r1208660408