workflow-context: update `repository` output
Previously was just using the `repository` key of the OIDC token, but
this actually points the the repo of the calling workflow.

This updates the output to pull out the repo from the `job_workflow_ref`,
which does point to the called reusable workflow.

(@tsibley already did this in his prototype but I totally missed it when
I was porting over the action.)