pathogen-repo-build: clone pathogen repo into top level working dir
Switch the directory organization within the workflow so that the
nextstrain/.github repo is in a subdir and the pathogen repo is now in
the top level working dir. This ensures that we don't have to worry
about prefixing the paths of the artifact paths.

We are cloning the nextstrain/.github repo into a hardcoded excluded
path¹ for AWS Batch jobs so that it does not get uploaded with the job.
I had explored other ways to try to exclude the extra support files
from the AWS Batch jobs, but they required some extra workarounds that
we don't want to deal with due to limitations of GitHub Actions.²

¹ https://github.com/nextstrain/cli/blob/6e36f6d7bdafcb984f62e2a9cc2a9cf9adf0be57/nextstrain/cli/runner/aws_batch/s3.py#L51-L66
² https://github.com/nextstrain/.github/pull/44#discussion_r1244500111