pathogen-repo-build: Replace CLI inputs with a single run input
Instead of recapitulating the `nextstrain build` options as inputs,
use a single run input for callers to provide the full command. This
seems like a simpler transition to be able to copy/paste the full
`nextstrain build` command already used locally to the automated
GH Action.

The expected `nextstrain build` commands are also simplified and will
no longer need to use the `--exec` option since the Nextstrain CLI now
has the `--env` option to pass environment variables to the build runtime.
This allows us to remove the write-envdir script and any references
to envdir.