pathogen-repo-build: remove hard-coded secrets
Allow callers to just pass secrets via `secrets: inherit` so that
we don't have to maintain a list of allowed secrets.

The only required/predefined secrets for setting up the aws-batch
runtime are added to the input description. All other secrets for builds
can have arbitrary names and can be passed to the build runtime
via the `--env` option.