pathogen-repo-build: Add failure notification
When GitHub Actions fail, they automatically send a notification to the
GitHub user who triggered the action or last edited the workflow cron
schedule¹. This additional Slack notification ensures that the whole
team is notified of failures. This has been our stopgap for monitoring
automated pathogen workflows since we have not set up more robust
monitoring dashboards. I expect this step to be removed once we migrate
to a different monitoring system outside of Slack.

¹ https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/notifications-for-workflow-runs