Rename "Testing" section to more generic "Checklist"
This section uses the checklist feature of GitHub, which renders in the
PR list view as "tasks" and a progress indicator to reflect how many
items are checked off.

Testing is just one of many things that should be done before a PR is
merged or even ready for review. Renaming to "Checklist" makes it more
obvious that arbitrary entries can be added and edited by the PR author,
which should improve the reviewing experience.