setup-nextstrain-cli: Run nextstrain update after runtime setup
There's a weird issue with the Conda runtime where the inital setup
uses an older version.¹ Do an update of the default runtime after setup
to ensure we are using the latest version.

We could remove this extra step after we rework `nextstrain setup` to
use the same logic as `nextstrain update`.

¹ https://github.com/nextstrain/monkeypox/issues/177