pathogen-repo-build: update AWS Batch job id matching
Previous pattern expected the AWS Batch job id to be in the example
re-attaching command that is only output by the Nextstrain CLI when
running the job with the `--detach` flag.

This checks for the AWS Batch job id in the stage output from the
CLI immediately after the job was successfully submitted to AWS Batch.¹
This output exists regardless of the detachment status.

¹ https://github.com/nextstrain/cli/blob/3c93e87c2ac6bdd1b6a913070f709067b29a6cc1/nextstrain/cli/runner/aws_batch/__init__.py#L258