pathogen-repo-build: Support assuming an AWS role via GitHub Actions' OIDC provider
The default role has a pretty limited set of permissions right now:
basically only those required to submit and interact with AWS Batch
jobs, but not, for example, permissions to upload to our production S3
buckets.  The permissions will probably need to be extended in the
future as repos/builds opt into this workflow and this role assumption.
The role's trust policy will also need to be extended to allow other
repos to assume it.

While the role is currently manually managed, it should really be
managed by a Terraform configuration (but maybe one separate from the
current nextstrain.org Terraform configurations?).

Related-to: <https://github.com/nextstrain/private/issues/22>