pathogen-repo-ci: Add `build-dir` input option
With the development of the pathogen-repo-template¹, we are reorganizing
pathogen repositories to move workflows into their own directories.
This change in monkeypox² revealed that the pathogen-repo-ci workflow no
longer works with this new file organization because it expects
`nextstrain build` to run with the root of the pathogen repository.

This commit adds support for this change by allowing users to define
the directory to use for the `nextstrain build` command.

Changes are backwards compatible as the default is set to `.` to point
to the root of the pathogen repository.

¹ https://github.com/nextstrain/pathogen-repo-template
² https://github.com/nextstrain/monkeypox/pull/198

Co-authored-by: Cornelius Roemer <cornelius.roemer@gmail.com>