pathogen-repo-ci: Call pathogen-repo-build internally
This allows the pathogen-repo-ci to continue to manage the complexity
related to testing multiple runtimes, but surfaces the flexibility of
the pathogen-repo-build inputs.

There are two major downsides to this approach:
1. We must hardcode the ref used for the pathogen-repo-build. We cannot
get around this with the workflow-context action because this is calling
a resuable workflow as a full job rather than an action step within a job.
2. We have to define additional permissions from the caller for both
the pathogen-repo-ci and the pathogen-repo-build with these updates
because if you specify the access for any of the scopes, all of those
that are not specified are set to none.