pathogen-repo-build: always generate AWS Batch Job ID & summary
Ran into a workflow run where the `run-build` step failed and the job
exited without generating the AWS Batch Job ID or summary.¹

This commit updates the condition to always try to generate the AWS
Batch ID as long as the runtime is `aws-batch` and the `build.log` file
exists. This then allows the workflow to always generate the AWS Batch
Job summary even if the build failed.

¹ https://github.com/nextstrain/forecasts-ncov/actions/runs/8203385071