pathogen-repo-build: Move `build.log` to ignored directory
The entire working directory gets zipped and uploaded to AWS Batch when
using the `aws-batch` runtime, which includes an early `build.log` file.
When the build completes and downloads the build outputs, the early
`build.log` overwrites the local `build.log` of the GH Action.
This results in a truncated file that does not include the
AWS_BATCH_JOB_ID and prevents the generation of the AWS Batch summary
and the re-attachment to the AWS Batch job in subsequent `wait` jobs.

This commit moves the hard-coded `build.log` file to an ignored directory
(the same directory that we are using for the nextstrain/.github repo¹).
This prevents the `build.log` from being uploaded with the build.

Previously tried other paths, but they all caused errors.
- "../build.log" - ".." pattern is not allowed for the `upload-artifact` action²
- "~/build.log" - `hashfiles` requires files to be in the `GITHUB_WORKSPACE`³

¹ https://github.com/nextstrain/.github/pull/44#discussion_r1253518104
² https://github.com/actions/upload-artifact/issues/176
³ https://docs.github.com/en/actions/learn-github-actions/expressions#hashfiles