pathogen-repo-build: support assuming AWS role for runtime permissions
Assumes repo-specific roles, `GitHubActionsRoleNextstrainRepo@zika` for
example, which are managed by the Terraform configuration in the
nextstrain/infra repo.  The repo here is always the _calling_
repository, regardless of if the "repo" input was provided to use
workflows from another place.

The runtime credentials are then saved in an envdir that is passed
to the build command via `NEXTSTRAIN_RUNTIME_ENVDIRS`.

Namespaced `NEXTSTRAIN_RUNTIME_ENVDIR` with `./git/nextstrain` as
suggested by @tsibley in review¹

¹ <https://github.com/nextstrain/.github/pull/81#discussion_r1602240011>

Related-to: <https://github.com/nextstrain/infra/issues/4>
Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>