actions/shellcheck: Hint to ShellCheck what shell flags are enabled via the shebang
This lets it properly detect when modes like "errexit" (i.e. `set -e`)
and "pipefail" are enabled via the shell invocation.  Resolves a false
warning in a Nextstrain CLI action that uses `pushd` without checking
exit code (because `-e` is in the shell invocation so checking is
unnecessary).

Disables a couple of warnings that pertain to the shebang, which are
fine to ignore because we're generating it just to hint ShellCheck.