pathogen-repo-ci: Don't run workflow build steps if they're not going to do anything
Instead of checking for whether to do anything or not _inside_ each
build step individually, move the check to the step's condition.  This
will make the logs a lot clearer and mean the GitHub workflow metadata
stays more in sync with reality.

I've used hashFiles() to check for file existence—it returns the empty
string, a falsey value, when there are no matching files—but we could
swap that out for using the output of a prior setup step that runs some
shell to determine what to run and what to skip.  Doing so might make
more sense if the conditional becomes more complicated or we want to do
more detailed reporting on _why_ steps were skipped or not.

With the changes, the internal-to-this-workflow action,
run-nextstrain-ci-build, is no longer that useful and so I've removed it
in favor of inlining things.  I think this is an improvement for
readability.

Related-to: <https://github.com/nextstrain/.github/pull/95#pullrequestreview-2113680284>