pathogen-repo-ci: Keep running even if one of the pathogen workflows fails
In testing it's helpful to report all test results even in the face of
an early failure to reduce the chances of playing whack-a-mole with
unreported failures once when you fix the first failure.  For example,
we want to know if the phylogenetic workflow fails even when the ingest
workflow run before it fails.