pathogen-repo-ci: Use status functions in conditionals instead of continue-on-error: true
While "continue-on-error: true" resulted in the behaviour we wanted, it
also meant the job step status indicators were misleading: even when an
ingest/phylogenetic/nextclade step failed, the UI showed a check mark
for them in the logs.  Instead of masking failures to keep going in the
face of errors, override the default success() conditional by specifying
our own conditions on a status expression.

Related-to: <https://github.com/nextstrain/.github/pull/98#issuecomment-2168846812>