actions/setup-debugger: Pause for 1 second after interrupting an exec…
…to allow /proc/$pid/cmdline to have contents.  It consistently seemed
to be empty (?) on GitHub Actions runners without this brief sleep.
This hacky solution worked in the moment, and so it's what I'm
committing since its tested.  There are likely better solutions that
don't rely on sleeping, like checking for the property we care about
(e.g. does /proc/$pid/cmdline have contents?), but I didn't try those
and so don't want to commit something that doesn't work.

Log the full argv for an easier time debugging the debugger next go
around.