Use a single build job
The number of steps that are shared between the two jobs now exceeds the
number of steps that differ. In this case, it seems more maintainable to
use one job with per-step conditions for environment setup.

This should also look nicer downstream - previously there was always one
job that was skipped since they are meant to be mutually exclusive.