pathogen-repo-build: update `wait-*` job conditionals
It's unclear to me why this change happened, but setting 
`continue-on-error: true` no longer masks the cancelled job's result as 
"successful". The job result is now "cancelled", so the `wait-*` jobs are 
skipped because the previous conditional depends on a "successful" upstream job.

This commit removes `continue-on-error: true` since it no longer serves it's 
purpose and updates the `wait-*` job conditional to run as long as the workflow
was not cancelled, the previous job did not fail, and the previous job 
conclusion was "cancelled" due to timeout. 

I am keeping the final `wait-conclusion` job as a marker of the workflow success 
in case we want to query the workflow status via the specific job status via 
the GitHub API.