pathogen-repo-build: run `wait-N` command with `timeout`
Run `wait-N` command with `timeout` to detach from AWS Batch before we hit the 
`job.timeout-minutes`. This allows us to avoid cancelling the job to avoid the 
"Cancelled" status for the whole workflow.¹

Setting timeout to 359m, which leaves a minute for the job to conclude before 
hitting the `job.timeout-minutes`.

¹ <https://github.com/nextstrain/.github/issues/129#issuecomment-2956350306>